Factor pattern analysis code out into its own header and source file.
