+++ explorer/interpreter/type_checker.cpp
@@ -17,6 +17,7 @@
 #include "explorer/common/error_builders.h"
 #include "explorer/interpreter/impl_scope.h"
 #include "explorer/interpreter/interpreter.h"
+#include "explorer/interpreter/pattern_analysis.h"
 #include "explorer/interpreter/value.h"
 #include "llvm/ADT/DenseSet.h"
 #include "llvm/ADT/StringExtras.h"
@@ -2919,382 +2920,6 @@ auto TypeChecker::TypeCheckPattern(
       return Success();
   }
 }
-
-// An abstracted view of a pattern or constant value (which we view as a
-// particular kind of pattern).
-class AbstractPattern {
- public:
-  enum Kind {
-    // A pattern that matches anything.
-    Wildcard,
-    // A pattern that matches a compound value with sub-patterns to match
-    // elements. A compound value is modeled as a discriminator name applied to
-    // a sequence of nested values: the alternative `Optional.Element(E)` has
-    // discriminator `Element` and nested value `E`, and the tuple `(A,B,C)`
-    // has an empty discriminator and nested values `A`, `B`, and `C`.
-    Compound,
-    // A pattern that matches a particular primitive value.
-    Primitive
-  };
-
-  // This is intentionally implicit to allow easy conversion from a container
-  // of `const Pattern*` to a container of `AbstractPattern`s.
-  // NOLINTNEXTLINE(google-explicit-constructor)
-  AbstractPattern(Nonnull<const Pattern*> pattern) { Set(pattern); }
-
-  AbstractPattern(Nonnull<const Value*> value, Nonnull<const Value*> type)
-      : value_(value), type_(type) {}
-
-  // Make a match-anything wildcard pattern.
-  static auto MakeWildcard() -> AbstractPattern {
-    return AbstractPattern(WildcardTag());
-  }
-
-  auto kind() const -> Kind {
-    if (auto* pattern = value_.dyn_cast<const Pattern*>()) {
-      return Compound;
-    }
-    if (auto* value = value_.dyn_cast<const Value*>()) {
-      if (isa<TupleValue, AlternativeValue>(value)) {
-        return Compound;
-      }
-      return Primitive;
-    }
-    CARBON_CHECK(value_.is<const WildcardTag*>());
-    return Wildcard;
-  }
-
-  auto discriminator() const -> std::string_view {
-    CARBON_CHECK(kind() == Compound);
-    if (auto* pattern = value_.dyn_cast<const Pattern*>()) {
-      if (auto* alt_pattern = dyn_cast<AlternativePattern>(pattern)) {
-        return alt_pattern->alternative_name();
-      }
-    } else if (auto* value = value_.dyn_cast<const Value*>()) {
-      if (auto* alt = dyn_cast<AlternativeValue>(value)) {
-        return alt->alt_name();
-      }
-    }
-    return {};
-  }
-
-  auto NumElements() const -> int {
-    if (auto* pattern = value_.dyn_cast<const Pattern*>()) {
-      if (auto* tuple_pattern = dyn_cast<TuplePattern>(pattern)) {
-        return tuple_pattern->fields().size();
-      } else if (isa<AlternativePattern>(pattern)) {
-        return 1;
-      }
-    } else if (auto* value = value_.dyn_cast<const Value*>()) {
-      if (auto* tuple = dyn_cast<TupleValue>(value)) {
-        return tuple->elements().size();
-      } else if (auto* alt = dyn_cast<AlternativeValue>(value)) {
-        return 1;
-      }
-    }
-    return 0;
-  }
-
-  void AppendElementsTo(std::vector<AbstractPattern>& out) const {
-    if (auto* pattern = value_.dyn_cast<const Pattern*>()) {
-      if (auto* tuple_pattern = dyn_cast<TuplePattern>(pattern)) {
-        auto fields = tuple_pattern->fields();
-        out.insert(out.end(), fields.begin(), fields.end());
-      } else if (auto* alt_pattern = dyn_cast<AlternativePattern>(pattern)) {
-        out.push_back(&alt_pattern->arguments());
-      }
-    } else if (auto* value = value_.dyn_cast<const Value*>()) {
-      if (auto* tuple = dyn_cast<TupleValue>(value)) {
-        auto* tuple_type = cast<TupleValue>(type_);
-        CARBON_CHECK(tuple->elements().size() == tuple_type->elements().size());
-        for (size_t i = 0; i != tuple->elements().size(); ++i) {
-          out.push_back(
-              AbstractPattern(tuple->elements()[i], tuple_type->elements()[i]));
-        }
-      } else if (auto* alt = dyn_cast<AlternativeValue>(value)) {
-        out.push_back(AbstractPattern(
-            &alt->argument(),
-            *cast<ChoiceType>(type_)->FindAlternative(alt->alt_name())));
-      }
-    }
-  }
-
-  auto value() const -> const Value& { return *value_.get<const Value*>(); }
-
-  auto type() const -> const Value& {
-    CARBON_CHECK(kind() == Compound);
-    return *type_;
-  }
-
- private:
-  // This is aligned so that we can use it in the `PointerUnion` below.
-  struct alignas(8) WildcardTag {};
-  AbstractPattern(WildcardTag)
-      : value_(static_cast<const WildcardTag*>(nullptr)), type_(nullptr) {}
-
-  void Set(Nonnull<const Pattern*> pattern) {
-    type_ = &pattern->static_type();
-    switch (pattern->kind()) {
-      case PatternKind::AddrPattern:
-      case PatternKind::AutoPattern:
-      case PatternKind::BindingPattern:
-      case PatternKind::GenericBinding:
-        value_ = static_cast<const WildcardTag*>(nullptr);
-        break;
-
-      case PatternKind::TuplePattern:
-      case PatternKind::AlternativePattern:
-        value_ = pattern;
-        break;
-
-      case PatternKind::ExpressionPattern:
-        value_ = &pattern->value();
-        break;
-
-      case PatternKind::VarPattern:
-        Set(&cast<VarPattern>(pattern)->pattern());
-        break;
-    }
-  }
-
-  // The underlying pattern: either a syntactic pattern, or a constant value,
-  // or a placeholder indicating that this is a wildcard pattern.
-  llvm::PointerUnion<Nonnull<const Pattern*>, Nonnull<const Value*>,
-                     const WildcardTag*>
-      value_;
-  // Values don't always know their types, so store the type here. We only
-  // really need it for the `const Value*` case but also store it for the
-  // `const Pattern*` case for convenience.
-  const Value* type_;
-};
-
-// A matrix of patterns, used for determining exhaustiveness and redundancy of
-// patterns in a match statement.
-//
-// See http://moscova.inria.fr/~maranget/papers/warn/index.html for details on
-// the algorithm used here.
-class PatternMatrix {
- public:
-  // Add a pattern vector row to this collection of pattern vectors.
-  void Add(std::vector<AbstractPattern> pattern_vector) {
-    CARBON_CHECK(matrix_.empty() || matrix_[0].size() == pattern_vector.size());
-    matrix_.push_back(std::move(pattern_vector));
-  }
-
-  // Is the given pattern vector redundant if it appears after the patterns in
-  // this matrix? That is, will it never match following the other patterns?
-  auto IsRedundant(llvm::ArrayRef<AbstractPattern> pattern) const -> bool {
-    return !IsUseful(pattern);
-  }
-
- private:
-  // The maximum number of times we will consider all alternatives when
-  // recursively expanding the pattern. Allowing this to happen an arbitrary
-  // number of times leads to exponential growth in the runtime of the
-  // algorithm.
-  static constexpr int MaxExponentialDepth = 8;
-
-  // Information about a constructor for a compound type.
-  struct DiscriminatorInfo {
-    // For an alternative, the name. Otherwise, empty.
-    std::string_view discriminator;
-    // The number of elements. For a tuple, the size. Always 1 for an
-    // alternative.
-    int size;
-  };
-
-  struct DiscriminatorSet {
-    std::vector<DiscriminatorInfo> found;
-    bool any_missing;
-  };
-
-  // Determine whether the given pattern vector is useful: that is, whether
-  // adding it to the matrix would allow any more values to be matched.
-  auto IsUseful(llvm::ArrayRef<AbstractPattern> pattern,
-                int max_exponential_depth = MaxExponentialDepth) const -> bool {
-    if (matrix_.empty()) {
-      return true;
-    }
-
-    CARBON_CHECK(pattern.size() == matrix_[0].size());
-    if (matrix_[0].empty()) {
-      return false;
-    }
-
-    switch (pattern[0].kind()) {
-      case AbstractPattern::Wildcard: {
-        auto discrim = FirstColumnDiscriminators();
-        // Check if we hit the depth limit. If so, we act as if the
-        // constructors present in this position are not exhaustive, that is,
-        // as if the type we're matching has some other constructor not
-        // corresponding to anything written in the pattern in this position.
-        // This can lead us to conclude that a pattern is useful if it is not,
-        // and that a set of patterns is not exhaustive when it is.
-        int new_depth =
-            max_exponential_depth - (discrim.found.size() > 1 ? 1 : 0);
-        if (!discrim.any_missing && new_depth >= 0) {
-          for (auto found : discrim.found) {
-            if (Specialize(found).IsUseful(*SpecializeRow(pattern, found),
-                                           new_depth)) {
-              return true;
-            }
-          }
-          return false;
-        }
-        return Default().IsUseful(pattern.slice(1), max_exponential_depth);
-      }
-
-      case AbstractPattern::Compound: {
-        DiscriminatorInfo discrim = {
-            .discriminator = pattern[0].discriminator(),
-            .size = pattern[0].NumElements()};
-        return Specialize(discrim).IsUseful(*SpecializeRow(pattern, discrim),
-                                            max_exponential_depth);
-      }
-
-      case AbstractPattern::Primitive: {
-        return Specialize(pattern[0].value())
-            .IsUseful(pattern.slice(1), max_exponential_depth);
-      }
-    }
-  }
-
-  // Find the discriminators used by the first column and check whether we
-  // found all of them.
-  auto FirstColumnDiscriminators() const -> DiscriminatorSet {
-    std::set<std::string_view> discrims;
-    std::optional<const ChoiceType*> choice;
-
-    for (auto& row : matrix_) {
-      CARBON_CHECK(!row.empty());
-      switch (row[0].kind()) {
-        case AbstractPattern::Wildcard:
-          continue;
-        case AbstractPattern::Compound: {
-          if (auto* tuple = dyn_cast<TupleValue>(&row[0].type())) {
-            // If we find a tuple match, we've found all constructors (there's
-            // only one!) and none were missing.
-            return {
-                .found = {{.discriminator = {},
-                           .size = static_cast<int>(tuple->elements().size())}},
-                .any_missing = false};
-          }
-          discrims.insert(row[0].discriminator());
-          choice = &cast<ChoiceType>(row[0].type());
-          break;
-        }
-        case AbstractPattern::Primitive: {
-          // TODO: We assume that primitive value matches are always incomplete
-          // for now. This might be reasonable for integer types, even i8, but
-          // is probably worth revisiting for `bool`.
-          return {.found = {}, .any_missing = true};
-        }
-      }
-    }
-
-    if (!choice || choice.value()->declaration().alternatives().size() !=
-                       discrims.size()) {
-      return {.found = {}, .any_missing = true};
-    }
-
-    DiscriminatorSet result = {.found = {}, .any_missing = false};
-    result.found.reserve(discrims.size());
-    for (auto s : discrims) {
-      result.found.push_back({.discriminator = s, .size = 1});
-    }
-    return result;
-  }
-
-  // Specialize the pattern vector `row` for the case that the first value
-  // matched uses `discriminator`.
-  static auto SpecializeRow(llvm::ArrayRef<AbstractPattern> row,
-                            DiscriminatorInfo discriminator)
-      -> std::optional<std::vector<AbstractPattern>> {
-    CARBON_CHECK(!row.empty());
-    std::vector<AbstractPattern> new_row;
-    switch (row[0].kind()) {
-      case AbstractPattern::Wildcard:
-        new_row.reserve(discriminator.size + row.size() - 1);
-        new_row.insert(new_row.end(), discriminator.size,
-                       AbstractPattern::MakeWildcard());
-        break;
-      case AbstractPattern::Compound: {
-        if (row[0].discriminator() != discriminator.discriminator) {
-          return std::nullopt;
-        }
-        CARBON_CHECK(static_cast<int>(row[0].NumElements()) ==
-                     discriminator.size);
-        new_row.reserve(discriminator.size + row.size() - 1);
-        row[0].AppendElementsTo(new_row);
-        break;
-      }
-      case AbstractPattern::Primitive:
-        // These cases should be rejected by the type checker.
-        llvm_unreachable("matched primitive against compound");
-    }
-    new_row.insert(new_row.end(), row.begin() + 1, row.end());
-    return std::move(new_row);
-  }
-
-  // Specialize the pattern matrix for the case that the first value matched
-  // uses `discriminator`, and its elements are matched.
-  auto Specialize(DiscriminatorInfo discriminator) const -> PatternMatrix {
-    PatternMatrix specialized;
-    for (auto& row : matrix_) {
-      // TODO: If we add support for "or" patterns, specialization might
-      // produce multiple rows here.
-      if (auto new_row = SpecializeRow(row, discriminator)) {
-        specialized.Add(std::move(new_row.value()));
-      }
-    }
-    return specialized;
-  }
-
-  // Specialize the pattern matrix for the case where the first value is known
-  // to be `value`, and is not matched.
-  auto Specialize(const Value& value) const -> PatternMatrix {
-    PatternMatrix specialized;
-    for (auto& row : matrix_) {
-      CARBON_CHECK(!row.empty());
-      switch (row[0].kind()) {
-        case AbstractPattern::Wildcard:
-          break;
-        case AbstractPattern::Compound:
-          llvm_unreachable("matched compound against primitive");
-        case AbstractPattern::Primitive:
-          // TODO: Use an equality context here?
-          if (!ValueEqual(&row[0].value(), &value, std::nullopt)) {
-            continue;
-          }
-          break;
-      }
-      specialized.Add(std::vector<AbstractPattern>(row.begin() + 1, row.end()));
-    }
-    return specialized;
-  }
-
-  // Specialize the pattern matrix for the case where the first value uses a
-  // discriminator matching none of the non-wildcard patterns.
-  auto Default() const -> PatternMatrix {
-    PatternMatrix default_matrix;
-    for (auto& row : matrix_) {
-      CARBON_CHECK(!row.empty());
-      switch (row[0].kind()) {
-        case AbstractPattern::Wildcard:
-          default_matrix.Add(
-              std::vector<AbstractPattern>(row.begin() + 1, row.end()));
-          break;
-        case AbstractPattern::Compound:
-        case AbstractPattern::Primitive:
-          break;
-      }
-    }
-    return default_matrix;
-  }
-
- private:
-  std::vector<std::vector<AbstractPattern>> matrix_;
-};
 
 auto TypeChecker::TypeCheckStmt(Nonnull<Statement*> s,
                                 const ImplScope& impl_scope)
